change cell numeric value type from long double to double, closes #201

+++ include/xlnt/utils/time.hpp
@@ -46,7 +46,7 @@ struct XLNT_API time
     /// The integer part of number will be ignored.
     /// 0.5 would return time(12, 0, 0, 0) or noon, halfway through the day.
     /// </summary>
-    static time from_number(long double number);
+    static time from_number(double number);
 
     /// <summary>
     /// Constructs a time object from an optional hour, minute, second, and microsecond.
@@ -62,7 +62,7 @@ struct XLNT_API time
     /// Returns a numeric representation of the time in the range 0-1 where the value
     /// is equal to the fraction of the day elapsed.
     /// </summary>
-    long double to_number() const;
+    double to_number() const;
 
     /// <summary>
     /// Returns true if this time is equivalent to comparand.
